requirements.txt
# Core ML/AI Libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
bitsandbytes==0.41.3

# MLflow
mlflow==2.9.2
mlflow-skinny==2.9.2

# API & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
pypdf==3.17.0
pypdf2==3.0.1
python-docx==1.1.0

# Guardrails
rebuff==0.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0

# AWS SDK (for deployment)
boto3==1.29.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

